+++ src/main/java/roomescape/controller/dto/TimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record TimeRequest(String startAt) {
